chore: Move core's postcss plugin to test helpers.
