+++ packages/@css-blocks/core/src/index.ts
@@ -2,7 +2,6 @@ export * from "./BlockTree";
 export * from "./errors";
 export * from "./SourceLocation";
 export * from "./configuration";
-export * from "./BlockParser";
 export * from "./BlockCompiler";
 export * from "./BlockParser";
 export * from "./errors";
@@ -10,7 +9,3 @@ export * from "./importing";
 export * from "./Analyzer";
 export * from "./TemplateRewriter";
 export * from "./configuration";
-
-import cssBlocks = require("./cssBlocks");
-// tslint:disable-next-line:no-default-export
-export default cssBlocks;
