+++ packages/@css-blocks/core/test/attribute-container-test.ts
@@ -3,10 +3,10 @@ import { assert } from "chai";
 import { suite, test } from "mocha-typescript";
 import { postcss } from "opticss";
 
-import cssBlocks = require("../src/cssBlocks");
 import { AttrValue } from "../src/index";
 
 import { BEMProcessor } from "./util/BEMProcessor";
+import cssBlocks = require("./util/postcss-helper");
 import { setupImporting } from "./util/setupImporting";
 
 @suite("Attribute container")
