+++ packages/@css-blocks/core/test/util/BEMProcessor.ts
@@ -2,7 +2,7 @@ import { postcss } from "opticss";
 import * as perfectionist from "perfectionist";
 
 import { Options } from "../../src/configuration";
-import cssBlocks = require("../../src/cssBlocks");
+import cssBlocks = require(".././util/postcss-helper");
 
 export class BEMProcessor {
   process(filename: string, contents: string, cssBlocksOpts?: Options) {
